requirement.txt
Django==1.9.2
django-debug-toolbar==1.4
httplib2==0.9.2
MySQL-python==1.2.5
Pillow==3.1.0
sqlparse==0.1.16
